requirements.txt
# ============================
# Security & Encryption
# ============================
cryptography==43.0.1
six>=1.16.0

# ============================
# Core Flask (Flask 3 compatible)
# ============================
Flask==3.0.3
Werkzeug>=3.0.0,<4
Jinja2==3.1.4
click>=8.1.7
itsdangerous>=2.1.2
blinker>=1.7.0
markupsafe>=2.1.5

# ============================
# Database
# ============================
Flask-MySQLdb==2.0.0
pymysql==1.1.0

# ============================
# Data & Processing
# ============================
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
pytz>=2024.1
python-dateutil>=2.9.0

# ============================
# Visualization & Dashboard
# ============================
dash==2.17.1
plotly==5.24.1

# ============================
# Python internal dependencies
# (IMPORTANT for PyInstaller / Electron)
# ============================
typing_extensions>=4.9.0,<5
importlib_metadata>=7.0,<8
zipp>=3.17,<4
packaging>=23.2,<25

# ============================
# Cloud & Storage
# ============================
cloudinary==1.41.0

# ============================
# Utilities
# ============================
python-dotenv==1.0.1
requests==2.32.3
num2words==0.5.13
weasyprint==61.2
waitress==3.0.1
flasgger==0.9.7.1